Adding new progress, revising def of codeword(W)
In matrixNew.txt, we just continue and get the final step, however, we observe that the definition of the W was not correct. Then at matrixNewRevised.txt, we changed the definition of codewords W = g*H and then tried to get the final step. We must then multiply the final result with Hadamard matrix at final step to get the final result!